optimize journal reading for multiple consumers by using a "wheel of entries"
The wheel of entries cycles through all entries as long as consumers are
available. Consumers can attach to the wheel whenever they want waiting until
the journal starts at the point that they are interested in and start consuming
from there on.

This allowed us to get rid of the "skipOverSnapshot" state in a process