Use Sha512 truncated to 160 bit as the default hash
The reason is that we otherwise spend so much space just saving the hashes.
160 bit should be plenty to avoid random collisions for reasonably sized
collections. It might not be enough against an active attacker.

This also needed some updates in the Server (where sha512 was hardcoded) and
in the repository code which can now find data also in the full sized hash
directory even if the default is different.